Migrate MCP build to Rolldown
